+++ src/components/doclist/views/IconsView.tsx
@@ -1,42 +1,16 @@
 'use client';
 
-import { useEffect, type CSSProperties } from 'react';
+import { useEffect } from 'react';
 import type { DocumentListDocument, IconSize } from '@/types/documents';
 import { DocumentTile } from './DocumentTile';
 import { useDocumentSelection } from '../dnd/DocumentSelectionContext';
+import { iconsGridStyle } from './iconsGrid';
 
 interface IconsViewProps {
   documents: DocumentListDocument[];
   iconSize: IconSize;
   onDeleteDoc: (doc: DocumentListDocument) => void;
   onMergeIntoFolder: (sources: DocumentListDocument[], target: DocumentListDocument) => void;
-}
-
-const TILE_WIDTH_PX: Record<IconSize, number> = {
-  sm: 112,
-  md: 136,
-  lg: 162,
-  xl: 192,
-};
-
-const SMALL_GRID_ITEM_COUNT = 3;
-
-function responsiveGridTemplate(iconSize: IconSize, itemCount: number): string {
-  const width = TILE_WIDTH_PX[iconSize];
-  if (itemCount <= SMALL_GRID_ITEM_COUNT) {
-    return `repeat(auto-fill, minmax(${width}px, ${width}px))`;
-  }
-  return `repeat(auto-fit, minmax(${width}px, 1fr))`;
-}
-
-const gridGap = '12px';
-
-function gridStyle(iconSize: IconSize, itemCount: number): CSSProperties {
-  return {
-    gridTemplateColumns: responsiveGridTemplate(iconSize, itemCount),
-    gap: gridGap,
-    justifyContent: itemCount <= SMALL_GRID_ITEM_COUNT ? 'start' : undefined,
-  };
 }
 
 export function IconsView({
@@ -61,7 +35,7 @@ export function IconsView({
       onClick={handleBackgroundClick}
       className="flex-1 min-h-0 overflow-y-auto p-3"
     >
-      <div className="grid" style={gridStyle(iconSize, documents.length)}>
+      <div className="grid" style={iconsGridStyle(iconSize, documents.length)}>
         {documents.map((doc) => (
           <DocumentTile
             key={`${doc.type}-${doc.id}`}
